requirements.txt
Django==1.11.5
oauthlib==2.0.0
requests==2.11.1
requests-oauthlib==0.6.2
PyJWT==1.5.3
gunicorn==19.6.0
